// src/app/shared/models/daily.forecast.ts
export interface DailyForecast {
  dayOfWeek: string;
  date: string;
  temp: {
    day: number;
    min: number;
    max: number;
    night: number;
    eve: number;
    morn: number;
  };
  feelsLike: {
    day: number;
    night: number;
    eve: number;
    morn: number;
  };
  pressure: number;
  humidity: number;
  windSpeed: number;
}